requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
Faker==8.12.1
fastapi==0.68.1
h11==0.12.0
idna==3.2
iniconfig==1.1.1
langdetect~=1.0.9
motor==3.0.0
numpy==1.21.2
odmantic==0.3.5
packaging==21.3
pandas==1.3.2
pluggy==1.0.0
prometheus-client==0.11.0
py==1.11.0
pydantic==1.8.2
pymongo==4.1.1
pyparsing==3.0.7
pytest==7.0.1
python-dateutil==2.8.2
python-dotenv==0.19.0
python-gitlab==2.10.1
pytz==2021.1
requests==2.26.0
requests-toolbelt==0.9.1
six==1.16.0
starlette==0.14.2
starlette-exporter==0.10.0
text-unidecode==1.3
tomli==2.0.1
tqdm==4.62.2
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0